Better import error messages (#6)
* Add *.png to .gitignore

This ignores the graph images that are generated when running tests.

* Print errors and don't run test for bad external imports

Previously, if a bad module was imported (e.g. if there is a syntax
error in the file), the parser would fail silently and attempt to run
the test. This would lead to cryptic and unhelpful error messages. Now,
the parser will return an error and prevent the test from being run if
any of the external imports are bad.

Test: From example/ run: PYTHONPATH=$PYTHONPATH:$PWD/.. python \
                         run_test.py example.test -a='ip="127.0.0.1"'

* Log when a transition is being run

Add a log message indicating that a transition is in the process of
running.