Improve error messaging. (#14)
* Improve error messaging.

As a first step in improved error messaging, setup a system
for pretty printing the parser errors. And make the error formatters
replaceable. This makes it easy to dump out computer readable formats
for use in build systems and human readable formats for manual use.

Issue: #2
Test: Add syntex error to example/example_base.stl and try to parse it.

* Update parser.py

Fix some bad formatting that happened

* Address comments in error_formatter.py

* Address comments for error_handler.py

* Addressed comments in parser.py

* Fix syntax error in error_formatter.py

* Fix indentation

* Fix copyright year

* Move ErrorInfo and ErrorPosition

* Move ErrorInfo and ErrorPostiion

* add missing import

* remove unused import